Move Socks5ProxyParams, Credentials into core
