+++ fee-provider/src/main/scala/org/bitcoins/feeprovider/BitcoinerLiveFeeRateProvider.scala
@@ -4,9 +4,9 @@ import akka.actor.ActorSystem
 import akka.http.scaladsl.model.Uri
 import org.bitcoins.commons.jsonmodels.wallet.BitcoinerLiveResult
 import org.bitcoins.commons.serializers.JsonSerializers._
+import org.bitcoins.core.api.tor.Socks5ProxyParams
 import org.bitcoins.core.wallet.fee.SatoshisPerVirtualByte
 import org.bitcoins.feeprovider.BitcoinerLiveFeeRateProvider.validMinutes
-import org.bitcoins.tor.Socks5ProxyParams
 import play.api.libs.json.{JsError, JsSuccess, Json}
 
 import scala.util.{Failure, Success, Try}
