+++ node/src/main/scala/org/bitcoins/node/util/PeerMessageSenderApi.scala
@@ -1,6 +1,7 @@
 package org.bitcoins.node.util
 
 import org.bitcoins.core.api.chain.{ChainApi, FilterSyncMarker}
+import org.bitcoins.core.api.node.Peer
 import org.bitcoins.core.number.Int32
 import org.bitcoins.core.p2p.{
   GetAddrMessage,
@@ -19,7 +20,6 @@ import org.bitcoins.core.protocol.transaction.Transaction
 import org.bitcoins.crypto.DoubleSha256DigestBE
 import org.bitcoins.node.config.NodeAppConfig
 import org.bitcoins.node.constant.NodeConstants
-import org.bitcoins.node.models.Peer
 import org.bitcoins.node.networking.peer.NodeState
 
 import scala.concurrent.{ExecutionContext, Future}
