+++ tor/src/main/scala/org/bitcoins/tor/client/TorClient.scala
@@ -2,10 +2,11 @@ package org.bitcoins.tor.client
 
 import grizzled.slf4j.Logging
 import org.bitcoins.commons.util.NativeProcessFactory
+import org.bitcoins.core.api.tor.Socks5ProxyParams
 import org.bitcoins.core.util.EnvUtil
 import org.bitcoins.tor.TorProtocolHandler._
 import org.bitcoins.tor.config.TorAppConfig
-import org.bitcoins.tor.{Socks5ProxyParams, TorParams}
+import org.bitcoins.tor.{TorParams}
 
 import java.io.{File, FileNotFoundException}
 import java.net.{InetAddress, InetSocketAddress}
